+++ packages/python/executors/prodkit-executor-database/src/prodkit_executor_database/__init__.py
@@ -3,4 +3,4 @@
 from .executor import ConstrainedDatabaseExecutor, DatabaseExecutorConfig
 
 __all__ = ["ConstrainedDatabaseExecutor", "DatabaseExecutorConfig"]
-__version__ = "0.9.0"
+__version__ = "0.9.1"