+++ packages/python/executors/prodkit-executor-http/src/prodkit_executor_http/__init__.py
@@ -3,4 +3,4 @@
 from .executor import ConstrainedHTTPExecutor, HTTPExecutorConfig
 
 __all__ = ("ConstrainedHTTPExecutor", "HTTPExecutorConfig")
-__version__ = "0.9.0"
+__version__ = "0.9.1"