+++ packages/python/prodkit-control-core/src/prodkit_control_core/__init__.py
@@ -8,4 +8,4 @@ from .exceptions import *
 from .ports import *
 from .reconciliation import *
 
-__version__ = "0.9.0"
+__version__ = "0.9.1"